+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `DeviceCoordinates` component
+
 ## [0.4.0] - 2020-02-27
 
 ### Added