+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- `DeviceCoordinates` component
+- `DeviceCoordinates` component.
 
 ## [0.4.0] - 2020-02-27