+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Default value for `display` prop in `PlaceDetails` component.
+- Prop `hiddenFields` in `AddressForm`.
 
 ### Changed
 - Update design of `LocationInput` component to match their spec.