+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.7.0] - 2020-04-22
 ### Added
 - Default value for `display` prop in `PlaceDetails` component.
 - Prop `hiddenFields` in `AddressForm` component.